Added support for different filetypes.
