+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "compress-create-react-app",
-  "version": "1.1.3",
+  "version": "1.1.4",
   "description": "An NPM package for adding post build compression to a create-react-app builds",
   "main": "index.js",
   "scripts": {
