+++ src/utils.js
@@ -22,6 +22,7 @@ const compressors = {
   },
 };
 
+
 const formatSizeUnits = (bytes) => {
   if (bytes >= 1073741824) {
     return (bytes / 1073741824).toFixed(2) + ' GB';
@@ -49,25 +50,6 @@ const getCombinedSize = (arrayOfFiles) => {
   return totalSize;
 };
 
-const getFiles = (
-  dirPath,
-  arrayOfFiles = [],
-  filetypes = ['.js', '.html', '.css']
-) => {
-  const files = fs.readdirSync(dirPath);
-  files.forEach((file) => {
-    if (fs.statSync(dirPath + '/' + file).isDirectory()) {
-      arrayOfFiles = getFiles(dirPath + '/' + file, arrayOfFiles, filetypes);
-    } else {
-      const filePath = path.join(dirPath, '/', file);
-      if (filetypes.includes(path.extname(filePath))) {
-        arrayOfFiles.push(filePath);
-      }
-    }
-  });
-  return arrayOfFiles;
-};
-
 const startCompressingFile = (file, algorithm) => {
   return new Promise(async (resolve, reject) => {
     try {
@@ -85,35 +67,13 @@ const startCompressingFile = (file, algorithm) => {
 const readFile = util.promisify(fs.readFile);
 const writeFile = util.promisify(fs.writeFile);
 
-const getBuildDirectory = () => {
-  const customBuildPathIndex =
-    process.argv.indexOf('--directory') !== -1
-      ? process.argv.indexOf('--directory') + 1
-      : process.argv.indexOf('-d') + 1;
 
-  const relativeBuildPath =
-    customBuildPathIndex === 0 ? '/build' : process.argv[customBuildPathIndex];
-
-  const absoluteBuildPath = path.join(appRoot, relativeBuildPath);
-  if (!fs.existsSync(absoluteBuildPath)) {
-    console.error(
-      '\x1b[31m%s\x1b[0m',
-      `No directory found at: ${path.join(appRoot, relativeBuildPath)}.`
-    );
-    console.log('Please provide the relative path from the project root');
-    process.exit(1);
-  }
-
-  return absoluteBuildPath;
-};
 
 module.exports = {
   compressors,
-  getFiles,
   getCombinedSize,
   formatSizeUnits,
   readFile,
   writeFile,
   startCompressingFile,
-  getBuildDirectory,
 };
